fixed pending errors

+++ src/app/modules/admin/manage-users/view-customer/view_customer.component.ts
@@ -79,7 +79,6 @@ export class ViewCustomerComponent implements OnInit {
           }
         });
       }
-
       getCategoryName(id){
         const res = this.categories.find( r=> r.id == 21);
         return res.category_name;
@@ -106,8 +105,6 @@ export class ViewCustomerComponent implements OnInit {
           if (resp && resp.success) {
             this.dashBoardData = resp.success.Data;
             this.dashboardInvestment.push(this.dashBoardData);
-            console.log(this.dashboardInvestment, 'HELLO');
-
           } else {
             this.dashBoardData = {number_of_pools: 0, investment_return: [], investment_report: []};
           }
@@ -115,19 +112,19 @@ export class ViewCustomerComponent implements OnInit {
         });
       }
 
-    updateUser(user, operation) {
+    updateUser(operation) {
         if (operation == 'enable') {
-          this.userService.activateUser(user).subscribe(resp => {
-            if (resp && resp.success) {
-              // alert(resp.success.Message)
-              // this.users[userIndex].email_is_verified=1
+          this.userService.activateUser(this.user).subscribe(resp => {
+            if(resp && resp.success) {
+               this.toastrService.success('User activated succesfully');
+               this.user.email_is_verified=1;
             }
           })
         }else{
-          this.userService.deactivateUser(user).subscribe(resp=>{
+          this.userService.deactivateUser(this.user).subscribe(resp=>{
             if(resp && resp.success){
-              // alert(resp.success.Message)
-              // this.users[userIndex].email_is_verified=0
+               this.toastrService.success('User deactivated succesfully');
+               this.user.email_is_verified=0;
             }
           });
         }
@@ -140,6 +137,7 @@ export class ViewCustomerComponent implements OnInit {
     }
 
     delete (users: User) {
+      if(confirm('Are you sure you want to delete user')){
         this.userService.deleteUser(users).subscribe(resp => {
           if (resp && resp.success) {
             this.toastrService.success('Details deleted succesfully');
@@ -149,26 +147,26 @@ export class ViewCustomerComponent implements OnInit {
           this.router.navigateByUrl('admin/manage-users');
         })
       }
-
-      calculateEstimate(returns, inv, expected_return_period, i) {
-        const estimate = (((returns * this.divisorFunc(expected_return_period, i)) - inv) / inv) * 100;
-        return Math.ceil(estimate);
-      }
-
-      divisorFunc (expected_return_period, i) {
-        if (this.userInvestment[i].expected_return_period === "Weekly") {
-          return 48;
-        } else if (this.userInvestment[i].expected_return_period === "Monthly") {
-          return 12;
-        }
-      };
-      addMonth(date: Date, month: number) {
-        const newDate = new Date(date);
-        const d = newDate.getDate();
-        newDate.setMonth(newDate.getMonth() + month);
-        if (newDate.getMonth() == 11) {
-            newDate.setDate(0);
-        }
-        return newDate;
     }
-  }
+    calculateEstimate(returns, inv, expected_return_period) {
+    const estimate = (((returns * this.divisorFunc(expected_return_period)) - inv) / inv) * 100;
+    return Math.ceil(estimate);
+    }
+
+    divisorFunc (expected_return_period) {
+    if ( expected_return_period === "Weekly") {
+        return 48;
+    } else if (expected_return_period === "Monthly") {
+        return 12;
+    }
+    };
+    addMonth(date: Date, month: number) {
+    const newDate = new Date(date);
+    const d = newDate.getDate();
+    newDate.setMonth(newDate.getMonth() + month);
+    if (newDate.getMonth() == 11) {
+        newDate.setDate(0);
+    }
+    return newDate;
+   }
+}
